fix: refactor single append and is_nil negation
Negation of is_nil is a bad practice once that is unclear what is the
point of the guard/function. You can always refactor to make pattern
macthing

Lists in Elixir are Linked Lists, so appending is a bad practice once is
has O(n). Instead you can prepedn the element to the list to get O(1)
operation