feat: add remaining typespecs
Although Elixir is a dynamic typed language, it's a good practice to add
typespecs to public function with the module attribute `@spec`. That not
change the semantic of the code but can help other developers to
understand the function context.

Also, Erlang/OTP has a tool called `dialyzer` that makes a static
analisys of your code, based on types. It can deduce and infer most of
the time but with those typespecs it's way easier to get better results.